% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ats_classification.R
\name{ats_classification}
\alias{ats_classification}
\title{ATS classification via Spirometry}
\usage{
ats_classification(data)
}
\arguments{
\item{data}{A data frame containing columns for fev1, fvc, fev1fvc, tlc, and their associated LLNs.}
}
\value{
The original data frame with an extra column for ATS classification.
}
\description{
Generates ATS classifications via spirometry and lung volume data.
}
\examples{
data <- data.frame(fev1,); ats_classification(data)

}
