% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion_capacity.R
\name{diffusion_normals}
\alias{diffusion_normals}
\title{Compute carbon monoxide diffusion capacity or transfer factor reference values for given demographics}
\usage{
diffusion_normals(data, SI.units = FALSE)
}
\arguments{
\item{data}{A data frame containing columns for sex ("M","F"),
age (in years, in the range 4.5-91) and height (in centimeters).}

\item{SI.units}{A boolean. Returns the reference values in SI units if TRUE
.  and Traditional units if FALSE.}
}
\value{
The original data frame with extra columns appended for each reference value computed.
}
\description{
`diffusion_normals()` computes ATS-compliant upper and lower normal limits
for carbon monoxide measured diffusion capacity and European equivalents
including DLCO (or TLCO), KCO, and VA.
}
\examples{
data <- data.frame(sex=c("M","F"),
                   age=c(30,5.1),
                   height=c(178,50))
diffusion_normals(data)

}
