% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirometry.R
\name{spirometry_normals}
\alias{spirometry_normals}
\title{Compute spirometry reference values for given demographics}
\usage{
spirometry_normals(data, year = 2012)
}
\arguments{
\item{data}{A data frame containing columns for sex ("M","F"),
race ("AfrAm","NEAsia","SEAsia","Other/mixed", "Caucasian"),
age (in years, in the range 3-95), and height (in centimeters).}

\item{year}{The year of GLI published equations to use in computing
spirometry  measures. Valid options are "2012" and "2022"}
}
\value{
The original data frame with extra columns appended for each reference value computed.
}
\description{
`spirometry_normals()` computes  ATS-compliant upper and lower normal limits
for common spirometry measures including FEV1, FVC, FEV1/FVC, FEF2575, and FEF75.
}
\examples{
data <- data.frame(sex=c("M","F"),
                   age=c(30.1,5.1),
                   height=c(178,50),
                   race=c("SEAsia","NEAsia"))
spirometry_normals(data)

}
