% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lung_volumes.R
\name{volume_normals}
\alias{volume_normals}
\title{Compute lung volume reference values for given demographics}
\usage{
volume_normals(data)
}
\arguments{
\item{data}{A data frame containing columns for sex ("M","F"),
age (in years, range 5 - 80) and height (in centimeters).}
}
\value{
The original data frame with extra columns appended for each reference value computed.
}
\description{
`volume_normals()` computes ATS-compliant upper and lower normal limits
for lung volume measures including FRC, TLC, RV, ERV, IC, and VC.
}
\examples{
data <- data.frame(sex=c("M","F"),
                   age=c(30,5.1),
                   height=c(178,50))
volume_normals(data)

}
\references{
Hall GL, Filipow N, Ruppel G, et al. Official ERS technical standard:
Global Lung Function Initiative reference values for static lung volumes
in individuals of European ancestry. Eur Respir J. 2021;57(3):2000289.
\doi{10.1183/13993003.00289-2020}.
}
