% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion_capacity.R
\name{diffusion_normals}
\alias{diffusion_normals}
\title{Compute carbon monoxide diffusion capacity or transfer factor reference values for given demographics}
\usage{
diffusion_normals(data, SI.units = FALSE)
}
\arguments{
\item{data}{A data frame containing columns for sex ("M","F"),
age (in years, in the range 5-90) and height (in centimeters).}

\item{SI.units}{A boolean. Returns the reference values in SI units if TRUE
.  and Traditional units if FALSE.}
}
\value{
The original data frame with extra columns appended for each reference value computed.
}
\description{
`diffusion_normals()` computes ATS-compliant upper and lower normal limits
for carbon monoxide measured diffusion capacity and European equivalents
including DLCO (or TLCO), KCO, and VA.
}
\examples{
data <- data.frame(sex=c("M","F"),
                   age=c(30,5.1),
                   height=c(178,50))
diffusion_normals(data)

}
\references{
Stanojevic S, Graham BL, Cooper BG, et al. Official ERS technical standards:
Global Lung Function Initiative reference values for the carbon monoxide
transfer factor for Caucasians. Eur Respir J. 2017;50(3):1700010.
\doi{10.1183/13993003.00010-2017}. (Author correction:
\doi{10.1183/13993003.50010-2017}, applied here.)
}
